+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `playVideo` function modified
 - `subSrc` renamed to `subtitleSrc`
 - `subType` renamed to `subtitleType`
+- `subFile` renamed to `subtitleFile`
 ## [0.9] - 2026-04-21
 ### Added
 - Video download button