+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `subSrc` renamed to `subtitleSrc`
 - `subType` renamed to `subtitleType`
 - `subFile` renamed to `subtitleFile`
+- `README.md` updated
 ## [0.9] - 2026-04-21
 ### Added
 - Video download button