+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `subType` renamed to `subtitleType`
 - `subFile` renamed to `subtitleFile`
 - `README.md` updated
+- `service-worker.js` modified
 ## [0.9] - 2026-04-21
 ### Added
 - Video download button