+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `subFile` renamed to `subtitleFile`
 - `README.md` updated
 - `service-worker.js` modified
+- Update bug fixed
 ## [0.9] - 2026-04-21
 ### Added
 - Video download button