+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.1] - 2026-05-24
 ### Added
 - Remove-all button
 ### Changed
@@ -109,7 +110,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Load subtitle from URL
 
 
-[Unreleased]: https://github.com/sepandhaghighi/neovid/compare/v1.0...main
+[Unreleased]: https://github.com/sepandhaghighi/neovid/compare/v1.1...main
+[1.1]: https://github.com/sepandhaghighi/neovid/compare/v1.0...v1.1
 [1.0]: https://github.com/sepandhaghighi/neovid/compare/v0.9...v1.0
 [0.9]: https://github.com/sepandhaghighi/neovid/compare/v0.8...v0.9
 [0.8]: https://github.com/sepandhaghighi/neovid/compare/v0.7...v0.8